refactor: restructure repository
