--- .ejs/cpp-class.ejs
@@ -1,26 +1,0 @@
-<%
-  const { CodeGenerator, CreateLogFile } = require('/home/krizsi90/Documents/Projects/UML_code_generator/.ejs/code-generator.js');
-  let header = CodeGenerator(element);
-
-  let publicElements = header.WritePublicElements();
-  let protectedElements = header.WriteProtectedElements();
-  let privateElements = header.WritePrivateElements();
--%>
-<%# Header guard start -%>
-#ifndef <%=header.WriteHeaderGuard()%>
-#define <%=header.WriteHeaderGuard()%>
-<%# Include headers -%>
-<%-header.WriteIncludes();%>
-
-<%# Class declaration -%>
-class <%=header.WriteName()%><%=header.WriteInheritance()%> {
-<%= publicElements != '' ? 'public:\n\t' + publicElements : ""-%>
-
-<%= protectedElements != '' ? 'protected:\n\t' + protectedElements : ""-%>
-
-<%= privateElements != '' ? 'private:\n\t' + privateElements : ""-%>
-};
-
-<%# Header guard end -%>
-#endif //<%=header.WriteHeaderGuard()%>
-<%CreateLogFile(header)-%>
