+++ src/enum_declaration.js
@@ -1,4 +1,4 @@
-const { getVisibility } = require('/home/krizsi90/Documents/Projects/UML_code_generator/.ejs/get_functions.js')
+const { getVisibility } = require('./get_functions.js')
 
 function collectUMLEnumerations(UMLelement) {
   let enumerations = [];
