+++ src/method_declaration.js
@@ -1,5 +1,4 @@
-const { getVisibility } = require('/home/krizsi90/Documents/Projects/UML_code_generator/.ejs/get_functions.js')
-const { getType } = require('/home/krizsi90/Documents/Projects/UML_code_generator/.ejs/get_functions.js')
+const { getVisibility, getType } = require('./get_functions.js')
 
 /**
  * declare all methods with the same type of visibility
